0x01: define `LIFOCache` caching system
It uses the LIFO algorithm for caching and adds the item with given `key`
in the cache. If the cache is full, it removes the last item inserted into
the cache to make room for the new item.
If the item already exists, it is replaced by the new item, and it
becomes the last item inserted into the cache.